+++ seleniumpm/webelements/clickable.py
@@ -1,4 +1,3 @@
-from seleniumpm.webelements.element import take_screenshot_on_element_error
 from seleniumpm.webelements.element import Element
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
@@ -9,7 +8,6 @@ class Clickable(Element):
     def __init__(self, driver, locator):
         super(Clickable, self).__init__(driver, locator)
 
-    @take_screenshot_on_element_error
     def wait_for_clickable(self, timeout=None):
         timeout = timeout if timeout is not None else self.element_timeout
         try:
@@ -24,7 +22,6 @@ class Clickable(Element):
             raise e
         return self
 
-    @take_screenshot_on_element_error
     def click(self, checkVisibility=False):
         """
         This is the standard selenium click() operation
@@ -35,7 +32,6 @@ class Clickable(Element):
             self.is_present_and_visible()
         self.get_webelement().click()
 
-    @take_screenshot_on_element_error
     def click_invisible(self):
         """
         This bypasses Selenium's enforcement on only allowing clicks on visible objects